+++ Realm/Tests/RealmTests.mm
@@ -33,7 +33,6 @@
 
 @interface RLMRealm ()
 + (BOOL)isCoreDebug;
-- (BOOL)compact;
 @end
 
 @interface RLMObjectSchema (Private)
@@ -1846,36 +1845,6 @@
     for (int i = 0; i < 9000; ++i) {
         [realm transactionWithBlock:^{}];
     }
-}
-
-- (void)testCompact
-{
-    RLMRealm *realm = self.realmWithTestPath;
-    NSString *uuid = [[NSUUID UUID] UUIDString];
-    NSUInteger count = 1000;
-    [realm transactionWithBlock:^{
-        [StringObject createInRealm:realm withValue:@[@"A"]];
-        for (NSUInteger i = 0; i < count; ++i) {
-            [StringObject createInRealm:realm withValue:@[uuid]];
-        }
-        [StringObject createInRealm:realm withValue:@[@"B"]];
-    }];
-    auto fileSize = ^(NSString *path) {
-        NSDictionary *attributes = [[NSFileManager defaultManager] attributesOfItemAtPath:path error:nil];
-        return [(NSNumber *)attributes[NSFileSize] unsignedLongLongValue];
-    };
-    unsigned long long fileSizeBefore = fileSize(realm.configuration.fileURL.path);
-    StringObject *object = [StringObject allObjectsInRealm:realm].firstObject;
-
-    XCTAssertTrue([realm compact]);
-
-    XCTAssertTrue(object.isInvalidated);
-    XCTAssertEqual([[StringObject allObjectsInRealm:realm] count], count + 2);
-    XCTAssertEqualObjects(@"A", [[StringObject allObjectsInRealm:realm].firstObject stringCol]);
-    XCTAssertEqualObjects(@"B", [[StringObject allObjectsInRealm:realm].lastObject stringCol]);
-
-    unsigned long long fileSizeAfter = fileSize(realm.configuration.fileURL.path);
-    XCTAssertGreaterThan(fileSizeBefore, fileSizeAfter);
 }
 
 - (NSArray *)pathsFor100Realms